Minor styling updates and whitespace removal.
