+++ django_analyses/filters/analysis.py
@@ -2,7 +2,7 @@
 Definition of an :class:`AnalysisFilter` for the
 :class:`~django_analyses.models.analysis.Analysis` model.
 """
-
+from django_analyses.filters.utils import DEFAULT_LOOKUP_CHOICES
 from django_analyses.models.analysis import Analysis
 from django_filters import rest_framework as filters
 
@@ -11,25 +11,14 @@ class AnalysisFilter(filters.FilterSet):
     """
     Provides useful filtering options for the
     :class:`~django_analyses.models.analysis.Analysis` model.
-
     """
 
-    title = filters.LookupChoiceFilter(
-        lookup_choices=[
-            ("contains", "Contains (case-sensitive)"),
-            ("icontains", "Contains (case-insensitive)"),
-            ("exact", "Exact"),
-        ]
-    )
+    title = filters.LookupChoiceFilter(lookup_choices=DEFAULT_LOOKUP_CHOICES)
     description = filters.LookupChoiceFilter(
-        lookup_choices=[
-            ("contains", "Contains (case-sensitive)"),
-            ("icontains", "Contains (case-insensitive)"),
-            ("exact", "Exact"),
-        ]
+        lookup_choices=DEFAULT_LOOKUP_CHOICES
     )
-    created = filters.DateTimeFromToRangeFilter("created")
+    created = filters.DateTimeFromToRangeFilter()
 
     class Meta:
         model = Analysis
-        fields = "id", "title", "description", "created", "category"
+        fields = "id", "category"
