+++ django_analyses/filters/utils.py
@@ -1,0 +1,8 @@
+"""
+Utilities for the :mod:`~django_analyses.filters` module.
+"""
+DEFAULT_LOOKUP_CHOICES = (
+    ("contains", "Contains (case-sensitive)"),
+    ("icontains", "Contains (case-insensitive)"),
+    ("exact", "Exact"),
+)
