+++ django_analyses/models/utils/run_status.py
@@ -1,0 +1,10 @@
+"""
+Definition of the :class:`RunStatus` :class:`Enum` subclass.
+"""
+from django_analyses.utils.choice_enum import ChoiceEnum
+
+
+class RunStatus(ChoiceEnum):
+    STARTED = "Started"
+    SUCCESS = "Success"
+    FAILURE = "Failure"
